tweaking sensor params

+++ hotpants-VCNL4000.py
@@ -113,7 +113,7 @@ def exit_handler():
     # uart.cleanup() # not yet supported?
 
 def mapVals(val, inMin, inMax, outMin, outMax):
-        toRet = outMin + ((outMax - outMin) * ((val - inMin) / (inMax - inMin)))
+        toRet = float(outMin + (float(outMax - outMin) * (float(val - inMin) / float(inMax - inMin))))
         # shall we clamp values?
         return toRet
